Separate helpers from artist meta component
